refactor: 각자 Domain Model에 배치
