+++ NewsResting/Domain/Models/NewsQuery.swift
@@ -6,8 +6,17 @@
 //
 
 import Foundation
+import CoreData
 
 struct NewsQuery: Equatable {
     let query: String
 }
 
+extension NewsQuery: EntityConvertable {
+    func toEntity(insertInto context: NSManagedObjectContext) -> NewsQueryEntity {
+        let entity = NewsQueryEntity(context: context)
+        entity.query = query
+        entity.createdAt = Date()
+        return entity
+    }
+}
